work/requirements.txt
absl-py==2.3.1
astunparse==1.6.3
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
Flask==2.2.5
flatbuffers==25.9.23
gast==0.4.0
google-auth==2.43.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.76.0
h5py==3.15.1
idna==3.11
ImageIO==2.37.2
itsdangerous==2.2.0
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.6
keras==2.10.0
Keras-Preprocessing==1.1.2
lazy_loader==0.4
libclang==18.1.1
Markdown==3.10
MarkupSafe==3.0.3
ml_dtypes==0.5.4
networkx==3.4.2
numpy==1.23.5
oauthlib==3.3.1
opt_einsum==3.4.0
packaging==25.0
pillow==12.0.0
protobuf==3.19.6
pyasn1==0.6.1
pyasn1_modules==0.4.2
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
scikit-image==0.24.0
scipy==1.15.3
six==1.17.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.1
tensorflow-cpu==2.12.0
tensorflow-estimator==2.10.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.2.0
tifffile==2025.5.10
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==1.14.2
